[breaking] Split the library to Core, DBus and Network libraries
This avoids users to not pull Qt dependencies into their applications
even if they don't use it just by linking against QCoro, e.g. we no
longer force all our users to link against QtDBus just because QCoro
supports it, but the application may not be using it. This will
become further more visible with Widgets supports.
