+++ QCoroConfig.cmake.in
@@ -2,11 +2,17 @@
 
 include(CMakeFindDependencyMacro)
 find_dependency(Qt@QT_VERSION_MAJOR@Core)
-find_dependency(Qt@QT_VERSION_MAJOR@DBus)
-find_dependency(Qt@QT_VERSION_MAJOR@Network)
-find_dependency(Qt@QT_VERSION_MAJOR@Widgets)
-find_dependency(Qt@QT_VERSION_MAJOR@Concurrent)
+
+set(QCORO_WITH_QTDBUS @QCORO_WITH_QTDBUS@)
+set(QCORO_WITH_QTNETWORK @QCORO_WITH_QTNETWORK@)
+
+if (QCORO_WITH_QTDBUS)
+    find_dependency(Qt@QT_VERSION_MAJOR@DBus)
+endif()
+if (QCORO_WITH_QTNETWORK)
+    find_dependency(Qt@QT_VERSION_MAJOR@Network)
+endif()
 
 include("${CMAKE_CURRENT_LIST_DIR}/QCoroTargets.cmake")
-
+# Custom macros
 include("${CMAKE_CURRENT_LIST_DIR}/QCoroMacros.cmake")
