+++ examples/CMakeLists.txt
@@ -1,6 +1,10 @@
 add_subdirectory(basics)
-add_subdirectory(dbus)
-add_subdirectory(network)
+if (QCORO_WITH_QTBUS)
+    add_subdirectory(dbus)
+endif()
+if (QCORO_WITH_QTNETWORK)
+    add_subdirectory(network)
+endif()
 if (QT_HAS_COMPAT_ABI)
     add_subdirectory(future)
 endif()
