+++ examples/chained/CMakeLists.txt
@@ -1,2 +1,5 @@
 add_executable(chained-example main.cpp)
-target_link_libraries(chained-example QCoro Qt${QT_VERSION_MAJOR}::Core)
+target_link_libraries(chained-example
+    QCoroCore
+    Qt${QT_VERSION_MAJOR}::Core
+)
