+++ examples/dbus/coroutine/CMakeLists.txt
@@ -4,7 +4,7 @@ set(dbustest_SRCS
 
 add_executable(dbustest-coro ${dbustest_SRCS})
 target_link_libraries(dbustest-coro
-    QCoro
+    QCoroDBus
     examples-dbus-common
     Threads::Threads
     Qt${QT_VERSION_MAJOR}::Core
