+++ examples/timer/CMakeLists.txt
@@ -1,2 +1,5 @@
 add_executable(timer-example main.cpp)
-target_link_libraries(timer-example QCoro Qt${QT_VERSION_MAJOR}::Core)
+target_link_libraries(timer-example
+    QCoroCore
+    Qt${QT_VERSION_MAJOR}::Core
+)
