+++ qcoro/core/iodevice.cpp
