+++ qcoro/core/qcoroiodevice.cpp
