+++ qcoro/core/qcoroprocess.cpp
