+++ qcoro/core/qcoroprocess.h
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "impl/waitoperationbase.h"
+#include "qcoro/waitoperationbase_p.h"
 #include "qcoroiodevice.h"
 
 #include <chrono>
