+++ qcoro/network/qcorotcpserver.h
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "impl/waitoperationbase.h"
+#include "qcoro/waitoperationbase_p.h"
 
 #include <QPointer>
