+++ tests/qdbuspendingreply.cpp
@@ -5,7 +5,8 @@
 #include "qcorodbustestinterface.h"
 #include "testdbusserver.h"
 #include "testobject.h"
-#include "qcoro/dbus.h"
+
+#include "qcoro/dbus/dbus.h"
 
 #include <QDBusConnection>
 #include <QDBusError>
